Prevent incorrect optimizations for window aggregates too
